Canonicalized URIs. Now they are of the form(s):
  wiki
  wiki?action=page

There is no such thing as "wiki?RecentChanges". It is now
"wiki?show=RecentChanges". This makes the code a bit easier to read, and
may help the search engines to actually spider the wiki.

All self-referencing wiki URIs are generated by a subroutine, so if I want
to try to change the URI form again it will be easier now. I was working on
that earlier tonight, but ran into trouble with how browsers generate the
queries for search forms.

Little steps now. Maybe bigger ones later.


git-svn-id: file:///home/david/tashtego/svn/wiki/trunk@360 f1bc7e2d-46f1-0310-96be-f91141564092