A slightly nicer factoring of new pages abstraction
This is a bit of a work-in-progress still, but I split up writing page
info in a nice way. You can write page attribs, or you can do put_page,
which first writes the page attribs, then deletes any attribs (from the
filesystem) that weren't in the hash of attribs that were written.