Canonicalise script hrefs if necessary
We turn "UseCanonicalName On" in the vhost config. However, this only
kicks in on _certain_ redirects (like mod_rewrite ones) that the server
generates. If I want to reliably send someone to a "canonical" domain, I
need to do it myself.

So, we now check if HTTP_HOST (the requested host, from Host: header)
and SERVER_NAME (which is _always_ set to ServerName if UseCanonicalName
is On) differ, create an _absolute_ script_href link that points to the
canonical host. Otherwise, we're already there, so just use regular
"root relative" URIs, like "/show/PageName".