Fixed interwiki linking! be careful with filepaths...
Because the actions are now in a subdir wrt to the other files, _any_
action that reads a file using a relative file path is relative to
actions/ subdir, and not the main wiki code dir.

The site-wide config.pl code was not being read in. Nor the intermap
file. So in particular interwiki links failed to render.