More careful generation of href URIs
There are some subtleties involved in generating correct hrefs for a,
link, and img elements.

First, some background. This wiki is designed so that the actions don't
have to be at the top (root) of the URI space - though the Apache config
that I have scripts to generate set it up that way. But in theory, we
could have a request URI like

  /path/to/actions/show/PageName

and with proper Apache setup, this would work fine. However, when show
script generates the page, it has to be careful about self-referential
links. It can't simply create links like

  <a href="/show/AnotherPage">AnotherPage</a>

since that'll point somewhere else in URI space than where the action
scripts live - but instead must generate

  <a href="/path/to/actions/show/AnotherPage">AnotherPage</a>

Likewise, if /show/PageName generates a page, links on that page can't
have the form

  <a href="show/AnotherPage">AnotherPage</a>

(note _lack_ of leading /) because that points to

  /show/PageName/show/AnotherPage.

So local links need to be "rooted" (ie, start with a slash). Basically,
they need to repeat the same path prefix as that used to execute the
script. (Generally this prefix is simply "/".)

But what about links off-site? These come in two flavors: <a> element
links that the reader can follow, and ones that are fetched
automatically - like <img /> or <link rel="stylesheet" />. In the second
case we use the URI unchanged.

But in the first case we generate a slightly odd href URI so that we can
see when the reader follows a link that _leaves_ the site:

  /path/prefix/out/http://external.page.com/foobar

The "out" is what is critical here. In our Apache config we match URIs
of that form, strip off the /path/prefix/out/ and redirect. But it
leaves a trace in our log file!

The factoring is still not perfect. But anywhere we make links for
images, stylesheets, and the like we call "rooted_href", which prepends
$pathprefix if the URI isn't already absolute (starting with an HTTP
scheme). And the function that makes <a> links, either prefixes the
passed-in URI with $pathprefix (for local links), or $pathprefix/out/
(for absolute links that point off-site).