Tweaked handling of site admin setting
It used to read from the _subdir_ (subdomain) directory, and used to
always append the main domain ($dom).

Now it reads from the main domain directory, $dir/../ - so all
subdomains share the same admin (maybe this is wrong?) - and runs
shell's eval, so it's possible to use vars such as $dom and $sub.

admin defaults to "webhamster@$dom" if the db/domain/admin file doesn't
exist.