Rename *.pl to *.perl
I think it reads better. When I was briefly exploring the Java world, I
remember noticing (and liking) that file extensions were spelled out in
full.

Does this mean that when I start using Python - which I plan to do for
this codebase - that I'll try to use .python as an extension? I doubt
it, of only because AFAIK Python doesn't understand that. I could be
wrong.