Git backend is Officially Live!!
Woohoo!

This was harder than I thought, partly due to quirks with Git when using
GIT_DIR and GIT_WORK_TREE to access a repo and file tree (resp) not in
the current directory. It turns out that you have to use _relative_
paths to files when talking to Git this way ... if in order to get to
the repo or tree (not sure which is the problem) you have to follow a
symlink!

I had to rethink a couple of things in pages.perl to get this right, and
it was surprisingly subtle, so I tried to carefully comment.

I also have no idea if this will work with concurrent access. While
futzing around I ran into an error message about a lock file on the
index, so that's maybe a good sign.

One other note about the way that I had to set things up to get it work.
I'm hoping/expecting to make some multi-page changes via sed, and so I
wanted shared command line and www user access to the pages and repo. I
had a bit of trouble setting this up - mostly because I forgot about the
umask. I had to set both my umask and the server's (if it loads
pages.perl) to 002.

I also did this:

  # chown -R www:staff .git newpages
  # chmod -R g+w .git newpages

(I'm on a Mac so group staff is in my list of groups.)

With the right umask, and these ownerships and permissions, sharing
between command line use and web use seamless.