Fixed edit collision code
It has been broken since I modularised the wiki, actually...

In this version it was always getting the edit area contents from the
existing page. In the collision case, you want to get the current
modtime of the file, but all other data should come from the POSTed
form.

That's what it does now.