"Improved" regular expressions for wikiwords and interwiki links
But I used the posix [[:unreadable:]] syntax for "clarity". Hah.

The interesting change is that interwiki query parts (the part following
the colon) cannot _end_ in a ")". Not sure this helps. My problem was
that I had an interwiki link inside parens; the RE ate the ")" and broke
the link.

I want to match () - they appear in Wikipedia links, eg - but what to do
if your query ends in ")" and you want to put it inside parens? I don't
know. Honestly.