Gets wiki code from $(pwd); db/ -> domains/
I forgot to change the reference to db/ to domains/ in the shell script.
This wreaks havoc.

Also, since the wiki code is included - not in a separate repo like
before - it is found simply by doing $(pwd) in the shell.

You have to set and export WEBHOME to your wwwroot as before though.

For testing, setting it to $(pwd)/www works well.