Added infrastructure for doing conditional GET (if-modified-since)
I've turned it off, because it interacts badly with the user experience.
Wikis work best when they dynamically (at the time of page rendering)
generate links, so they can be coloured (or elided entirely) based on
the existence of the linked-to page.

Until I have some kind of dependency network (linksto/from) I'll leave
this off. I may need it - or at the dependency stuff - when I start
thinking about how to push pages to Blogger...