$iconhref -> $iconpage. Icon links to pages only
On second thought I decided that really I just want to be able to set
the page that the icon points to. If $iconpage is unset, it goes to
$defaultpage, as before.

One nice change here is that the $editable variable is set before
loading config.perl, so it is possible to have different $iconpage
settings for different versions of the site.