Improved Installed_Successfully page
It's much more readable now. The full docroot is mentioned only once, in
a <pre> section, and mentions of images/ and files/ directories, and how
to make links to files in them, are also broken out into <pre>'s, and
use a much shorter, relative path (basically the domain name).